@@ -3,16 +3,17 @@
   "description": "netrc auth plugin for Octokit",
   "license": "MIT",
   "version": "0.0.0-semantically-released",
+  "type": "module",
   "engines": {
-    "node": "^14.15 || >=16"
+    "node": "^18.17"
   },
   "author": "Matt Travi <npm@travi.org> (https://matt.travi.org/)",
   "repository": "travi/octokit-auth-netrc",
   "bugs": "https://github.com/travi/octokit-auth-netrc/issues",
   "homepage": "https://npm.im/octokit-auth-netrc",
   "runkitExampleFilename": "./example.js",
-  "main": "lib/index.cjs.js",
-  "module": "lib/index.es.js",
+  "main": "./lib/index.js",
+  "exports": "./lib/index.js",
   "sideEffects": false,
   "scripts": {
     "pretest": "run-s build",
@@ -30,17 +31,18 @@
     "watch": "run-s 'build:js -- --watch'",
     "prepack": "run-s build",
     "test:unit": "cross-env NODE_ENV=test c8 run-s test:unit:base",
-    "test:unit:base": "DEBUG=any mocha 'src/**/*-test.js'",
+    "test:unit:base": "DEBUG=any vitest run",
     "lint:peer": "npm ls >/dev/null",
     "lint:gherkin": "gherkin-lint",
-    "pretest:integration": "preview",
+    "pretest:integration": "run-s build",
     "test:integration": "run-s 'test:integration:base -- --profile noWip'",
-    "test:integration:base": "DEBUG=any cucumber-js test/integration --profile base",
+    "test:integration:base": "DEBUG=any NODE_OPTIONS=--no-experimental-fetch cucumber-js test/integration",
     "test:integration:debug": "DEBUG=test run-s test:integration",
     "test:integration:wip": "run-s 'test:integration:base -- --profile wip'",
     "test:integration:focus": "run-s 'test:integration:base -- --profile focus'",
     "prepare": "husky install",
-    "lint:engines": "ls-engines"
+    "lint:engines": "ls-engines",
+    "lint:publish": "publint --strict"
   },
   "files": [
     "example.js",
@@ -51,42 +53,39 @@
     "provenance": true
   },
   "dependencies": {
-    "@octokit/auth-token": "^3.0.0",
+    "@octokit/auth-token": "^4.0.0",
     "netrc": "0.1.4"
   },
   "devDependencies": {
-    "@babel/register": "7.22.15",
-    "@cucumber/cucumber": "9.6.0",
+    "@cucumber/cucumber": "10.0.1",
     "@octokit/core": "4.2.4",
     "@travi/any": "3.0.1",
-    "@travi/babel-preset": "3.0.113",
     "@travi/eslint-config": "5.0.2",
     "@travi/eslint-config-cucumber": "1.0.6",
-    "@travi/eslint-config-mocha": "1.0.21",
     "ban-sensitive-files": "1.10.1",
     "c8": "8.0.1",
     "chai": "4.3.10",
-    "commitlint-config-travi": "1.4.39",
+    "commitlint-config-travi": "1.4.40",
     "cross-env": "7.0.3",
     "cz-conventional-changelog": "3.3.0",
     "debug": "4.3.4",
     "gherkin-lint": "4.2.2",
     "http-status-codes": "2.3.0",
     "husky": "8.0.3",
+    "jest-when": "3.6.0",
     "lockfile-lint": "4.12.1",
     "ls-engines": "0.9.0",
-    "mocha": "10.2.0",
     "mock-fs": "5.2.0",
     "nock": "13.3.6",
     "npm-run-all2": "6.1.1",
-    "package-preview": "4.0.0",
-    "remark-cli": "11.0.0",
-    "remark-preset-lint-travi": "6.0.1",
+    "publint": "0.2.5",
+    "remark-cli": "12.0.0",
+    "remark-preset-lint-travi": "6.0.3",
     "remark-toc": "9.0.0",
     "remark-usage": "11.0.0",
     "rimraf": "5.0.5",
-    "rollup": "2.79.1",
+    "rollup": "4.1.4",
     "rollup-plugin-auto-external": "2.0.0",
-    "sinon": "17.0.0"
+    "vitest": "0.34.6"
   }
 }
