@@ -1,6 +1,5 @@
 import {createTokenAuth} from '@octokit/auth-token';
-
-import netrc from '../thirdparty-wrappers/netrc.js';
+import netrc from 'netrc';
 
 export default function ({domain = 'api.github.com'} = {}) {
   const githubCredentials = netrc()[domain];
